Fix snprintf-calls on 32 bit architectures
- use %lli and cast from time_t to long long because the size time_t can be 32 bit or 64 bit even on 32 bit architectures
- use %zd for arguments with the type size_t because the size of size_t depends on the architecture

+++ src/db_sqlite3.c
@@ -917,8 +917,8 @@ db_sqlite3_post_insert_into_postings(server_cb_inf *inf, char *message_id,
 
 	sqlite3_snprintf((2*len) - 1, sql_cmd,
 		"insert into postings (msgid, date, author, newsgroups, subject, "
-		"lines, header) values ('%q', '%li', '%q', '%q', '%q', '%i', '%q');",
-		message_id, ltime, from, ngstrpb, subj, linecount, add_to_hdr);
+		"lines, header) values ('%q', '%lli', '%q', '%q', '%q', '%i', '%q');",
+		message_id, (long long) ltime, from, ngstrpb, subj, linecount, add_to_hdr);
 
 	sqlite3_secexec(inf, sql_cmd, NULL, 0);
 	sqlite3_free(sql_cmd);
